fix: move export to csv dialog in backend-ai-session-list to backend-ai-session-view (#815)
